GPII-1897: Added node-specific loader for gpii.pouch (available in gpii.pouch.node).
