+++ tests/js/pouchdb-component-tests/html/index.html
@@ -12,6 +12,7 @@
         <script type="text/javascript" src="../../../../node_modules/infusion/src/framework/core/js/FluidIoC.js"></script>
         <script type="text/javascript" src="../../../../node_modules/infusion/src/framework/core/js/DataBinding.js"></script>
         <script type="text/javascript" src="../../../../node_modules/infusion/src/framework/core/js/FluidView.js"></script>
+        <script type="text/javascript" src="../../../../node_modules/infusion/src/framework/core/js/FluidPromises.js"></script>
         <script type="text/javascript" src="../../../../node_modules/infusion/src/framework/core/js/ModelTransformation.js"></script>
         <script type="text/javascript" src="../../../../node_modules/infusion/src/framework/core/js/ModelTransformationTransforms.js"></script>
         <script type="text/javascript" src="../../../../node_modules/infusion/src/framework/enhancement/js/ContextAwareness.js"></script>
@@ -33,7 +34,10 @@
 
         <!-- This package's code to be tested-->
         <script type="text/javascript" src="../../../../src/test/common-helpers.js"></script>
-        <script type="text/javascript" src="../../../../src/js/pouchdb.js"></script>
+        <script type="text/javascript" src="../../../../src/js/pouchdb-common.js"></script>
+
+        <!-- required for use with Testem -->
+        <script type="text/javascript" src="/testem.js"></script>
 
         <link rel="stylesheet" href="../../../../node_modules/node-jqunit/node_modules/infusion/tests/lib/qunit/css/qunit.css"/>
     </head>
@@ -45,6 +49,9 @@
         <ol id="qunit-tests"></ol>
 
         <!-- our QUnit tests -->
-        <script type="text/javascript" src="../index.js"></script>
+        <script type="text/javascript" src="../pouchdb-component-common-tests.js"></script>
+        <script type="text/javascript">
+            fluid.test.runTests("gpii.tests.pouchdb.component.common.environment");
+        </script>
     </body>
 </html>
