+++ tests/js/pouchdb-component-tests/index.js
@@ -1,325 +1,69 @@
-// Browser and Node shared fixtures for the `gpii.pouch` PouchDB component tests.
+/*
+
+    Run the common tests with the node component type.  Also test unique functionality of this grade.
+
+ */
 /* eslint-env node */
 "use strict";
-var fluid  = fluid || require("infusion");
-var gpii   = fluid.registerNamespace("gpii");
-var jqUnit = jqUnit || require("node-jqunit");
+var fluid = require("infusion");
 
-if (!gpii.pouch) {
-    require("../../../")
-    gpii.pouch.loadTestingSupport();
-}
+require("./pouchdb-component-common-tests");
 
-fluid.registerNamespace("gpii.tests.pouchdb.component");
-
-// A simple function to test support for "map" functions using the query method.
-/* globals emit */
-gpii.tests.pouchdb.component.map = function (doc) {
-    if (doc.color) {
-        emit(doc.color, doc);
-    }
-};
-
-// A simple function to test the support for "reduce" functions using the query method.
-gpii.tests.pouchdb.component.reduce = function (keys, values, rereduce) {
-    var colorSummary = {};
-    if (rereduce) {
-        values.forEach(function (toReduce) {
-            Object.keys(toReduce).forEach(function (key) {
-                var subtotal = toReduce[key];
-                colorSummary[key] = colorSummary[key] ? colorSummary[key] + subtotal : subtotal;
-            });
-        });
-    }
-    else {
-        keys.forEach(function (entry) {
-            var color = entry[0];
-            colorSummary[color] = colorSummary[color] ? colorSummary[color] + 1 : 1;
-        });
-    }
-    return colorSummary; // { red: 1, yellow: 2 }
-};
-
-fluid.defaults("gpii.tests.pouchdb.component.caseHolder", {
-    gradeNames: ["gpii.test.express.caseHolder.base"],
-    inputs: {
-        bulkData: [
-            {
-                "_id": "one",
-                "saying": "is the loneliest number"
-            },
-            {
-                "_id": "two",
-                "saying": "are required to tango"
-            }
-        ],
-        mapReduce: [
-            {
-                "color": "red",
-                "name": "strawberry"
-            },
-            {
-                "color": "yellow",
-                "name": "banana"
-            },
-            {
-                "color": "red",
-                "name": "cherry"
-            },
-            {
-                "color": "blue",
-                "name": "blueberry"
-            },
-            {
-                name: "colorless glass noodles"
-            }
-        ]
-    },
-    expectedResponses: {
-        bulkCreate: [{ ok: true, id: "one"}, { ok: true, id: "two"}]
-    },
-    // Recreate the database before each test.
-    sequenceStart: [{
-        func: "{testEnvironment}.events.constructFixtures.fire"
-    }],
-    // Destroy the database after each test.
-    sequenceEnd:   [
-        {
-            func: "{testEnvironment}.pouchDb.destroyPouch"
-        },
-        {
-            event:    "{testEnvironment}.pouchDb.events.onDestroyPouchComplete",
-            listener: "jqUnit.assertLeftHand",
-            args:     ["The database should be destroyed on test completion...", { ok: true}, "{arguments}.0"]
-        }
-    ],
+fluid.defaults("gpii.tests.pouchdb.component.node.caseHolder", {
+    gradeNames: ["gpii.tests.pouchdb.component.common.caseHolder.base"],
     rawModules: [{
-        name: "Tests for `gpii.pouch` component...",
+        name: "Data-loading tests for the `gpii.pouch.node` component...",
         type: "test",
         tests: [
             {
-                name: "Test the `info` method...",
+                name: "Confirm that onDataLoaded is fired if we have no data...",
                 sequence: [
                     {
-                        func: "{testEnvironment}.pouchDb.info"
+                        func: "{testEnvironment}.pouchDb.loadData",
+                        args: []
                     },
                     {
-                        event:    "{testEnvironment}.pouchDb.events.onInfoComplete",
-                        listener: "jqUnit.assertLeftHand",
-                        args:     ["The database info should be as expected...", { db_name: "test", doc_count: 0, update_seq: 0}, "{arguments}.0"]
-                    }
-                ]
-            },
-            {
-                name: "Test the `POST` method (with an ID)...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.post",
-                        args: [{ _id: "foo"}]
-                    },
-                    {
-                        event: "{testEnvironment}.pouchDb.events.onPostComplete",
-                        listener: "jqUnit.assertLeftHand",
-                        args: ["The POST should have been successful...", { ok: true, id: "foo"}, "{arguments}.0"]
-                    }
-                ]
-            },
-            {
-                name: "Test the `GET` method (with a non-existent ID)...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.get",
-                        args: [{ _id: "notGonnaFindIt"}]
-                    },
-                    {
-                        event: "{testEnvironment}.pouchDb.events.onError",
+                        event:    "{testEnvironment}.pouchDb.events.onDataLoaded",
                         listener: "jqUnit.assert",
-                        args: ["An error should be thrown if a record cannot be found..."]
+                        args:    ["onDataLoaded should be fired even if we have no data..."]
                     }
                 ]
             },
             {
-                name: "Test the `PUT` method (with an ID)...",
+                name: "Test loadData with a single value...",
                 sequence: [
                     {
-                        func: "{testEnvironment}.pouchDb.post",
-                        args: [{ _id: "bar"}]
+                        func: "{testEnvironment}.pouchDb.loadData",
+                        args: ["%gpii-pouchdb/tests/data/rgb.json"]
                     },
                     {
-                        event: "{testEnvironment}.pouchDb.events.onPostComplete",
-                        listener: "jqUnit.assertLeftHand",
-                        args: ["The PUT should have been successful...", { ok: true, id: "bar"}, "{arguments}.0"]
-                    }
-                ]
-            },
-            {
-                name: "Test the `POST` method (without an ID)...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.post",
-                        args: [{ "onions": "raw" }]
-                    },
-                    {
-                        event: "{testEnvironment}.pouchDb.events.onPostComplete",
-                        listener: "{testEnvironment}.pouchDb.get",
-                        args: ["{arguments}.0.id"]
-                    },
-                    {
-                        event: "{testEnvironment}.pouchDb.events.onGetComplete",
-                        listener: "jqUnit.assertLeftHand",
-                        args: ["The record should have been POSTed...", { "onions": "raw"}, "{arguments}.0"]
-                    }
-                ]
-            },
-            {
-                name: "Test the `bulkDocs` function...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.bulkDocs",
-                        args: ["{that}.options.inputs.bulkData"]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onBulkDocsComplete",
-                        listener: "gpii.test.pouchdb.recordsAreEquivalent",
-                        args:     ["The bulk creation of documents should have been successful...", "{that}.options.expectedResponses.bulkCreate", "{arguments}.0"]
-                    }
-                ]
-            },
-            {
-                name: "Test the `allDocs` function...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.bulkDocs",
-                        args: ["{that}.options.inputs.bulkData"]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onBulkDocsComplete",
-                        listener: "{testEnvironment}.pouchDb.allDocs",
-                        args:     []
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onAllDocsComplete",
-                        listener: "jqUnit.assertLeftHand",
-                        args:     ["The bulk retrieval of all documents should return the correct number of records...", { offset: 0, total_rows: 2 }, "{arguments}.0"]
-                    }
-                ]
-            },
-            {
-                name: "Test the `bulkGet` function...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.bulkDocs",
-                        args: ["{that}.options.inputs.bulkData"]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onBulkDocsComplete",
-                        listener: "{testEnvironment}.pouchDb.bulkGet",
-                        args:     [{ docs: [{ id: "one" }, { id: "two"}]}]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onBulkGetComplete",
-                        listener: "gpii.test.pouchdb.recordsAreEquivalent",
-                        args:     ["a bulk get of all documents should return the correct content...", [{ id: "one" }, { id: "two"}], "{arguments}.0.results"]
-                    }
-                ]
-            },
-            {
-                name: "Test the `remove` function...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.bulkDocs",
-                        args: ["{that}.options.inputs.bulkData"]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onBulkDocsComplete",
-                        listener: "{testEnvironment}.pouchDb.get",
-                        args:     ["two"]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onGetComplete",
-                        listener: "{testEnvironment}.pouchDb.remove",
-                        args:     ["{arguments}.0"]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onRemoveComplete",
+                        event:    "{testEnvironment}.pouchDb.events.onDataLoaded",
                         listener: "{testEnvironment}.pouchDb.info",
                         args:     []
                     },
                     {
                         event:    "{testEnvironment}.pouchDb.events.onInfoComplete",
                         listener: "jqUnit.assertLeftHand",
-                        args:     ["The database should now only contain one record...", { db_name: "test", doc_count: 1 }, "{arguments}.0"]
+                        args:     ["The database should have been loaded correctly...", { db_name: "test", doc_count: 4}, "{arguments}.0"]
                     }
                 ]
             },
             {
-                name: "Test the `remove` method (with a non-existent ID)...",
+                name: "Test loadData with an array...",
                 sequence: [
                     {
-                        func: "{testEnvironment}.pouchDb.remove",
-                        args: [{ _id: "notGonnaFindIt"}]
+                        func: "{testEnvironment}.pouchDb.loadData",
+                        args: [["%gpii-pouchdb/tests/data/rgb.json", "%gpii-pouchdb/tests/data/supplemental.json"]]
                     },
                     {
-                        event: "{testEnvironment}.pouchDb.events.onError",
-                        listener: "jqUnit.assert",
-                        args: ["An error should be thrown if a record cannot be found..."]
-                    }
-                ]
-            },
-            {
-                name: "Testing `query` handling (passing a single map function)...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.bulkDocs",
-                        args: ["{that}.options.inputs.mapReduce"]
+                        event:    "{testEnvironment}.pouchDb.events.onDataLoaded",
+                        listener: "{testEnvironment}.pouchDb.info",
+                        args:     []
                     },
                     {
-                        event:    "{testEnvironment}.pouchDb.events.onBulkDocsComplete",
-                        listener: "{testEnvironment}.pouchDb.query",
-                        args:     [gpii.tests.pouchdb.component.map]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onQueryComplete",
-                        listener: "jqUnit.assertEquals",
-                        args:     ["The view should return the right number of records...", 4, "{arguments}.0.rows.length" ]
-                    }
-                ]
-            },
-            {
-                name: "Testing `query` handling (passing a map function as part of a JSON document)...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.bulkDocs",
-                        args: ["{that}.options.inputs.mapReduce"]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onBulkDocsComplete",
-                        listener: "{testEnvironment}.pouchDb.query",
-                        args:     [{ map: gpii.tests.pouchdb.component.map}]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onQueryComplete",
-                        listener: "jqUnit.assertEquals",
-                        args:     ["The view should return the right number of records...", 4, "{arguments}.0.rows.length" ]
-                    }
-                ]
-            },
-            {
-                name: "Testing `query` handling (passing both map and reduce function)...",
-                sequence: [
-                    {
-                        func: "{testEnvironment}.pouchDb.bulkDocs",
-                        args: ["{that}.options.inputs.mapReduce"]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onBulkDocsComplete",
-                        listener: "{testEnvironment}.pouchDb.query",
-                        args:     [{ map: gpii.tests.pouchdb.component.map, reduce: gpii.tests.pouchdb.component.reduce}]
-                    },
-                    {
-                        event:    "{testEnvironment}.pouchDb.events.onQueryComplete",
-                        listener: "jqUnit.assertDeepEq",
-                        args:     ["The reduced data should be correct...", { red: 2, yellow: 1, blue: 1 }, "{arguments}.0.rows.0.value" ]
+                        event:    "{testEnvironment}.pouchDb.events.onInfoComplete",
+                        listener: "jqUnit.assertLeftHand",
+                        args:     ["The contents of both data files should have been loaded...", { db_name: "test", doc_count: 5}, "{arguments}.0"]
                     }
                 ]
             }
@@ -327,32 +71,78 @@ fluid.defaults("gpii.tests.pouchdb.component.caseHolder", {
     }]
 });
 
-fluid.defaults("gpii.tests.pouchdb.component.environment", {
-    gradeNames: ["fluid.test.testEnvironment"],
-    events: {
-        constructFixtures: null
-    },
+fluid.defaults("gpii.tests.pouchdb.component.node.loadDataOnStartup.caseHolder", {
+    gradeNames: ["gpii.tests.pouchdb.component.common.caseHolder.base"],
+    rawModules: [{
+        name: "Test data loading on startup for the `gpii.pouch.node` component...",
+        type: "test",
+        tests: [
+            {
+                name: "Confirm that data is loaded on startup...",
+                sequence: [
+                    /*
+                        I know the component will already exist, but if I don't have this step, I get errors like:
+
+                        Message: Assertion failure (see console.log for expanded message): Error resolving path segment
+                        "pouchDb" of path pouchDb.events.onDataLoaded since component with record ,[object Object], has
+                        annotation "createOnEvent" - this very likely represents an implementation error.
+
+                        TODO:  Discuss with Antranig.
+
+                     */
+                    {
+                        func: "fluid.identity"
+                    },
+                    {
+                        event:    "{testEnvironment}.pouchDb.events.onDataLoaded",
+                        listener: "{testEnvironment}.pouchDb.info",
+                        args:    []
+                    },
+                    {
+                        event:    "{testEnvironment}.pouchDb.events.onInfoComplete",
+                        listener: "jqUnit.assertLeftHand",
+                        args:     ["The database should have been loaded correctly...", { db_name: "test", doc_count: 3}, "{arguments}.0"]
+                    }
+                ]
+            }
+        ]
+    }]
+});
+
+
+fluid.defaults("gpii.tests.pouchdb.component.node.common.environment", {
+    gradeNames: ["gpii.tests.pouchdb.component.common.environment"],
     components: {
         pouchDb: {
-            type: "gpii.pouch",
-            createOnEvent: "constructFixtures",
-            options: {
-                dbOptions: {
-                    name: "test"
-                }
-            }
-        },
-        caseHolder: {
-            type: "gpii.tests.pouchdb.component.caseHolder"
+            type: "gpii.pouch.node.base"
         }
     }
 });
 
-fluid.test.runTests("gpii.tests.pouchdb.component.environment");
+fluid.defaults("gpii.tests.pouchdb.component.node.environment", {
+    gradeNames: ["gpii.tests.pouchdb.component.node.common.environment"],
+    components: {
+        caseHolder: {
+            type: "gpii.tests.pouchdb.component.node.caseHolder"
+        }
+    }
+});
 
-/*
+fluid.defaults("gpii.tests.pouchdb.component.node.loadDataOnStartup.environment", {
+    gradeNames: ["gpii.tests.pouchdb.component.node.common.environment"],
+    components: {
+        pouchDb: {
+            type: "gpii.pouch.node",
+            options: {
+                dbPaths: ["%gpii-pouchdb/tests/data/data.json"]
+            }
+        },
+        caseHolder: {
+            type: "gpii.tests.pouchdb.component.node.loadDataOnStartup.caseHolder"
+        }
+    }
+});
 
-404 testing for all record methods
-
- // TODO: The change listeners appear to fire too quickly, we need another way to test them.
- */
+fluid.test.runTests("gpii.tests.pouchdb.component.node.common.environment");
+fluid.test.runTests("gpii.tests.pouchdb.component.node.environment");
+fluid.test.runTests("gpii.tests.pouchdb.component.node.loadDataOnStartup.environment");
